added getUser decorators, added added yaml config files

+++ src/tasks/tasks.service.ts
@@ -16,14 +16,14 @@ export class TasksService {
     ){}
 
 
-    public getTasks(filterDto:GetTasksFilterDto):Promise<Task[]> {
-        return this._taskRepository.getTasks(filterDto);
+    public getTasks(filterDto:GetTasksFilterDto, user:User):Promise<Task[]> {
+        return this._taskRepository.getTasks(filterDto, user);
     }
 
-    public async getTaskById(id:number):Promise<Task>{
+    public async getTaskById(id:number, user:User):Promise<Task>{
         // https://typeorm.delightful.studio/classes/_repository_repository_.repository.html
-        const found = await this._taskRepository.findOne(id);
-        if(!found) throw new NotFoundException(`Task with ID ${id} not found`);
+        const found = await this._taskRepository.findOne({where: {id, userId:user.id}});
+        if(!found) throw new NotFoundException(`Task with ID ${id} not found for the user`);
         return found;
     }
 
@@ -31,14 +31,14 @@ export class TasksService {
         return await this._taskRepository.createTask(createTaskDto, user);
     }
 
-    public async deleteTaskById(id:number):Promise<void> {
-        const result = await this._taskRepository.delete(id);
-        console.log(result);
+    public async deleteTaskById(id:number, user:User):Promise<void> {
+        const result = await this._taskRepository.delete({id, userId: user.id});
         if(result.affected === 0) throw new NotFoundException(`Task with ID ${id} not found`);
     }
 
-     public async udpateTaskStatus(id:number, status:TASK_STATUS): Promise<Task> {
-        const task = await this.getTaskById(id);
+     public async udpateTaskStatus(
+         id:number, status:TASK_STATUS, user:User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
